requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
email-validator
requests
httpx
python-dotenv
google-generativeai
numpy
pandas
Pillow
sentence-transformers
chromadb
langchain
langchain-core
langchain-openai
langchain-google-genai
langchain-community
torch
pymysql